IA-32 fixes. Use kernel addresses instead of physical addresses in map_page_to_frame(). Physical addresses are supposed to only be exported to mm hardware. Because of this fix, userspace is functional again.
Remap EGA videoram to (0x80000000 + videoram) and change the ega driver to work with the new address.

Minor cosmetics through out the code.
Changes in linker scripts.

+++ src/main/main.c
@@ -55,6 +55,8 @@
 #include <arch.h>
 #include <arch/faddr.h>
 
+#include <typedefs.h>
+
 char *project = "SPARTAN kernel";
 char *copyright = "Copyright (C) 2001-2005 Jakub Jermar\nCopyright (C) 2005 HelenOS project";
 
@@ -67,8 +69,8 @@ context_t ctx;
  * appropriate sizes and addresses.
  */
 __address hardcoded_load_address = 0;
-__u32 hardcoded_ktext_size = 0;
-__u32 hardcoded_kdata_size = 0;
+size_t hardcoded_ktext_size = 0;
+size_t hardcoded_kdata_size = 0;
 
 void main_bsp(void);
 void main_ap(void);
